requirements.txt
aiogram==3.10.0
openai==1.35.0
yt-dlp
python-dotenv